Connected To Firebase

+++ app/src/main/java/com/liez/plan_plant/SensorCahaya.kt
@@ -1,5 +1,5 @@
 package com.liez.plan_plant
 
 data class SensorCahaya(
-    val nilai : Float?
+    val nilai : Double?
 )
